linux-fresh: Add Clang 12 from llvm.org
Adds the Clang compiler from apt.llvm.org to linux-fresh. The intention
is to run two instances of linux-fresh when checking CI on yuzu, one
compiles with GCC and the other with Clang.

This includes a workaround in Conan where the default configuration file
is not set up to permit using Clang 12. settings.yml is there to allow
using Clang 12.