mingw: Remove glibc workaround
ccache requires a newer version of glibc than could be provided with the
Azure workaround. Revert 87c73e5b5d2dee97bef8d1ef6c17c610872a22a7 so
that ccache will work.

This may not work with Azure. If that is the case, we will need to
revert ccache to an earlier build.